chore: move APIs into ClientDevicesAuthServiceApi
