+++ src/main/java/com/aws/greengrass/device/ClientDevicesAuthService.java
@@ -16,10 +16,7 @@ import com.aws.greengrass.dependency.ImplementsService;
 import com.aws.greengrass.deployment.DeviceConfiguration;
 import com.aws.greengrass.device.configuration.GroupConfiguration;
 import com.aws.greengrass.device.configuration.GroupManager;
-import com.aws.greengrass.device.exception.AuthenticationException;
-import com.aws.greengrass.device.exception.AuthorizationException;
 import com.aws.greengrass.device.exception.CloudServiceInteractionException;
-import com.aws.greengrass.device.iot.IotAuthClient;
 import com.aws.greengrass.device.session.MqttSessionFactory;
 import com.aws.greengrass.device.session.SessionConfig;
 import com.aws.greengrass.device.session.SessionCreator;
@@ -49,8 +46,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ThreadPoolExecutor;
@@ -86,12 +81,10 @@ public class ClientDevicesAuthService extends PluginService {
 
     private final GreengrassServiceClientFactory clientFactory;
 
+    private final ClientDevicesAuthServiceApi clientDevicesAuthServiceApi;
     private final DeviceConfiguration deviceConfiguration;
     private final AuthorizationHandler authorizationHandler;
     private final GreengrassCoreIPCService greengrassCoreIPCService;
-    private final IotAuthClient iotAuthClient;
-    private final SessionManager sessionManager;
-    private final DeviceAuthClient deviceAuthClient;
     // Limit the queue size before we start rejecting requests
     private static final int DEFAULT_CLOUD_CALL_QUEUE_SIZE = 100;
     private static final int DEFAULT_THREAD_POOL_SIZE = 1;
@@ -103,17 +96,16 @@ public class ClientDevicesAuthService extends PluginService {
     /**
      * Constructor.
      *
-     * @param topics                   Root Configuration topic for this service
-     * @param groupManager             Group configuration management
-     * @param certificateManager       Certificate management
-     * @param clientFactory            Greengrass cloud service client factory
-     * @param deviceConfiguration      core device configuration
-     * @param authorizationHandler     authorization handler for IPC calls
-     * @param greengrassCoreIPCService core IPC service
-     * @param mqttSessionFactory       session factory to handling mqtt credentials
-     * @param iotAuthClient            iot auth client
-     * @param sessionManager           session manager
-     * @param deviceAuthClient         device auth client
+     * @param topics                      Root Configuration topic for this service
+     * @param groupManager                Group configuration management
+     * @param certificateManager          Certificate management
+     * @param clientFactory               Greengrass cloud service client factory
+     * @param deviceConfiguration         core device configuration
+     * @param authorizationHandler        authorization handler for IPC calls
+     * @param greengrassCoreIPCService    core IPC service
+     * @param mqttSessionFactory          session factory to handling mqtt credentials
+     * @param sessionManager              session manager
+     * @param clientDevicesAuthServiceApi client devices service api handle
      */
     @SuppressWarnings("PMD.ExcessiveParameterList")
     @Inject
@@ -123,9 +115,8 @@ public class ClientDevicesAuthService extends PluginService {
                                     AuthorizationHandler authorizationHandler,
                                     GreengrassCoreIPCService greengrassCoreIPCService,
                                     MqttSessionFactory mqttSessionFactory,
-                                    IotAuthClient iotAuthClient,
                                     SessionManager sessionManager,
-                                    DeviceAuthClient deviceAuthClient) {
+                                    ClientDevicesAuthServiceApi clientDevicesAuthServiceApi) {
         super(topics);
         cloudCallQueueSize = DEFAULT_CLOUD_CALL_QUEUE_SIZE;
         cloudCallQueueSize = getValidCloudCallQueueSize(topics);
@@ -133,15 +124,13 @@ public class ClientDevicesAuthService extends PluginService {
                 DEFAULT_THREAD_POOL_SIZE, 60, TimeUnit.SECONDS,
                 new ResizableLinkedBlockingQueue<>(cloudCallQueueSize));
         cloudCallThreadPool.allowCoreThreadTimeOut(true); // act as a cached threadpool
+        this.clientDevicesAuthServiceApi = clientDevicesAuthServiceApi;
         this.groupManager = groupManager;
         this.certificateManager = certificateManager;
         this.clientFactory = clientFactory;
         this.deviceConfiguration = deviceConfiguration;
         this.authorizationHandler = authorizationHandler;
         this.greengrassCoreIPCService = greengrassCoreIPCService;
-        this.iotAuthClient = iotAuthClient;
-        this.sessionManager = sessionManager;
-        this.deviceAuthClient = deviceAuthClient;
         SessionCreator.registerSessionFactory("mqtt", mqttSessionFactory);
         certificateManager.updateCertificatesConfiguration(new CertificatesConfig(this.getConfig()));
         sessionManager.setSessionConfig(new SessionConfig(this.getConfig()));
@@ -247,13 +236,14 @@ public class ClientDevicesAuthService extends PluginService {
         greengrassCoreIPCService.setSubscribeToCertificateUpdatesHandler(context ->
                 new SubscribeToCertificateUpdatesOperationHandler(context, certificateManager, authorizationHandler));
         greengrassCoreIPCService.setVerifyClientDeviceIdentityHandler(context ->
-                new VerifyClientDeviceIdentityOperationHandler(context, this, authorizationHandler,
-                        cloudCallThreadPool));
+                new VerifyClientDeviceIdentityOperationHandler(context, clientDevicesAuthServiceApi,
+                        authorizationHandler, cloudCallThreadPool));
         greengrassCoreIPCService.setGetClientDeviceAuthTokenHandler(context ->
-                new GetClientDeviceAuthTokenOperationHandler(context, this, authorizationHandler,
+                new GetClientDeviceAuthTokenOperationHandler(context, clientDevicesAuthServiceApi, authorizationHandler,
                         cloudCallThreadPool));
         greengrassCoreIPCService.setAuthorizeClientDeviceActionHandler(context ->
-                new AuthorizeClientDeviceActionOperationHandler(context, this, authorizationHandler));
+                new AuthorizeClientDeviceActionOperationHandler(context, clientDevicesAuthServiceApi,
+                        authorizationHandler));
     }
 
     public CertificateManager getCertificateManager() {
@@ -345,43 +335,5 @@ public class ClientDevicesAuthService extends PluginService {
             throw new CloudServiceInteractionException(
                     String.format("Failed to put core %s CA certificates to cloud", thingName), e);
         }
-    }
-
-    /**
-     * Verify client device identity.
-     * @param certificatePem PEM encoded client certificate.
-     * @return True if the provided client certificate is trusted.
-     */
-    public boolean verifyClientDeviceIdentity(String certificatePem) {
-        // Allow internal clients to verify their identities
-        if (deviceAuthClient.isGreengrassComponent(certificatePem)) {
-            return true;
-        } else {
-            Optional<String> certificateId = iotAuthClient.getActiveCertificateId(certificatePem);
-            return certificateId.isPresent();
-        }
-    }
-
-    /**
-     * Get client auth token.
-     * @param credentialType    Type of client credentials
-     * @param deviceCredentials Client credential map
-     * @return client auth token to be used for future authorization requests.
-     * @throws AuthenticationException if unable to authenticate client credentials
-     */
-    public String getClientDeviceAuthToken(String credentialType, Map<String, String> deviceCredentials)
-        throws AuthenticationException {
-        return sessionManager.createSession(credentialType, deviceCredentials);
-    }
-
-    /**
-     * Authorize client action.
-     * @param authorizationRequest Authorization request, including auth token, operation, and resource
-     * @return true if the client action is allowed
-     * @throws AuthorizationException if the client action is not allowed
-     */
-    public boolean authorizeClientDeviceAction(AuthorizationRequest authorizationRequest)
-            throws AuthorizationException {
-        return deviceAuthClient.canDevicePerform(authorizationRequest);
     }
 }
