+++ src/main/java/com/aws/greengrass/device/ClientDevicesAuthServiceApi.java
@@ -1,0 +1,75 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.aws.greengrass.device;
+
+import com.aws.greengrass.device.exception.AuthenticationException;
+import com.aws.greengrass.device.exception.AuthorizationException;
+import com.aws.greengrass.device.iot.IotAuthClient;
+import com.aws.greengrass.device.session.SessionManager;
+
+import java.util.Map;
+import java.util.Optional;
+import javax.inject.Inject;
+
+public class ClientDevicesAuthServiceApi {
+    private final IotAuthClient iotAuthClient;
+    private final SessionManager sessionManager;
+    private final DeviceAuthClient deviceAuthClient;
+
+    /**
+     * Constructor.
+     *
+     * @param iotAuthClient    iot auth client
+     * @param sessionManager   session manager
+     * @param deviceAuthClient device auth client
+     */
+    @Inject
+    public ClientDevicesAuthServiceApi(IotAuthClient iotAuthClient,
+                                       SessionManager sessionManager,
+                                       DeviceAuthClient deviceAuthClient) {
+        this.iotAuthClient = iotAuthClient;
+        this.sessionManager = sessionManager;
+        this.deviceAuthClient = deviceAuthClient;
+    }
+
+    /**
+     * Verify client device identity.
+     * @param certificatePem PEM encoded client certificate.
+     * @return True if the provided client certificate is trusted.
+     */
+    public boolean verifyClientDeviceIdentity(String certificatePem) {
+        // Allow internal clients to verify their identities
+        if (deviceAuthClient.isGreengrassComponent(certificatePem)) {
+            return true;
+        } else {
+            Optional<String> certificateId = iotAuthClient.getActiveCertificateId(certificatePem);
+            return certificateId.isPresent();
+        }
+    }
+
+    /**
+     * Get client auth token.
+     * @param credentialType    Type of client credentials
+     * @param deviceCredentials Client credential map
+     * @return client auth token to be used for future authorization requests.
+     * @throws AuthenticationException if unable to authenticate client credentials
+     */
+    public String getClientDeviceAuthToken(String credentialType, Map<String, String> deviceCredentials)
+        throws AuthenticationException {
+        return sessionManager.createSession(credentialType, deviceCredentials);
+    }
+
+    /**
+     * Authorize client action.
+     * @param authorizationRequest Authorization request, including auth token, operation, and resource
+     * @return true if the client action is allowed
+     * @throws AuthorizationException if the client action is not allowed
+     */
+    public boolean authorizeClientDeviceAction(AuthorizationRequest authorizationRequest)
+            throws AuthorizationException {
+        return deviceAuthClient.canDevicePerform(authorizationRequest);
+    }
+}
