+++ src/main/java/com/aws/greengrass/ipc/VerifyClientDeviceIdentityOperationHandler.java
@@ -10,6 +10,7 @@ import com.aws.greengrass.authorization.AuthorizationHandler;
 import com.aws.greengrass.authorization.Permission;
 import com.aws.greengrass.authorization.exceptions.AuthorizationException;
 import com.aws.greengrass.device.ClientDevicesAuthService;
+import com.aws.greengrass.device.ClientDevicesAuthServiceApi;
 import com.aws.greengrass.logging.api.Logger;
 import com.aws.greengrass.logging.impl.LogManager;
 import com.aws.greengrass.util.Utils;
@@ -37,7 +38,7 @@ public class VerifyClientDeviceIdentityOperationHandler
     private static final String UNAUTHORIZED_ERROR = "Not Authorized";
     private static final String NO_DEVICE_CREDENTIAL_ERROR = "Client device credential is required";
     private static final String NO_DEVICE_CERTIFICATE_ERROR = "Client device certificate is required";
-    private final ClientDevicesAuthService clientDevicesAuthService;
+    private final ClientDevicesAuthServiceApi clientDevicesAuthServiceApi;
     private final String serviceName;
     private final AuthorizationHandler authorizationHandler;
     private final ExecutorService cloudCallThreadPool;
@@ -45,20 +46,20 @@ public class VerifyClientDeviceIdentityOperationHandler
     /**
      * Constructor.
      *
-     * @param context                  operation continuation handler
-     * @param clientDevicesAuthService client devices auth service handle
-     * @param authorizationHandler     authorization handler
-     * @param cloudCallThreadPool      executor to run the call to the cloud asynchronously
+     * @param context                     operation continuation handler
+     * @param clientDevicesAuthServiceApi client devices auth service handle
+     * @param authorizationHandler        authorization handler
+     * @param cloudCallThreadPool         executor to run the call to the cloud asynchronously
      */
     public VerifyClientDeviceIdentityOperationHandler(
-            OperationContinuationHandlerContext context, ClientDevicesAuthService clientDevicesAuthService,
+            OperationContinuationHandlerContext context, ClientDevicesAuthServiceApi clientDevicesAuthServiceApi,
             AuthorizationHandler authorizationHandler, ExecutorService cloudCallThreadPool) {
 
         super(context);
-        this.clientDevicesAuthService = clientDevicesAuthService;
-        serviceName = context.getAuthenticationData().getIdentityLabel();
+        this.clientDevicesAuthServiceApi = clientDevicesAuthServiceApi;
         this.authorizationHandler = authorizationHandler;
         this.cloudCallThreadPool = cloudCallThreadPool;
+        serviceName = context.getAuthenticationData().getIdentityLabel();
     }
 
     @Override
@@ -88,7 +89,7 @@ public class VerifyClientDeviceIdentityOperationHandler
             String certificate = getCertificateFromCredential(request.getCredential());
             try {
                 VerifyClientDeviceIdentityResponse response = new VerifyClientDeviceIdentityResponse();
-                response.withIsValidClientDevice(clientDevicesAuthService.verifyClientDeviceIdentity(certificate));
+                response.withIsValidClientDevice(clientDevicesAuthServiceApi.verifyClientDeviceIdentity(certificate));
                 return response;
             } catch (Exception e) {
                 logger.atError().cause(e).log("Unable to verify client device identity");
